4,5問目完了

+++ src/lesson3/Recipe.php
@@ -5,7 +5,7 @@ namespace prphp\lesson3;
 class Recipe
 {
     private string $food;
-    private int $calorie;
+    private float $calorie;
 
     /**
      * @param string $food
